WEBAPP-231: Fix tests and move getLanguageChangePath to app
