+++ src/modules/app/getLanguageChangePath.js
@@ -1,9 +1,9 @@
 import CategoriesMapModel from '../endpoint/models/CategoriesMapModel'
-import { EXTRAS_ROUTE, getExtraPath } from '../app/routes/extras'
-import { DISCLAIMER_ROUTE, getDisclaimerPath } from '../app/routes/disclaimer'
-import { EVENTS_ROUTE, getEventPath } from '../app/routes/events'
-import { getSearchPath, SEARCH_ROUTE } from '../app/routes/search'
-import { CATEGORIES_ROUTE, getCategoryPath } from '../app/routes/categories'
+import { EXTRAS_ROUTE, getExtraPath } from './routes/extras'
+import { DISCLAIMER_ROUTE, getDisclaimerPath } from './routes/disclaimer'
+import { EVENTS_ROUTE, getEventPath } from './routes/events'
+import { getSearchPath, SEARCH_ROUTE } from './routes/search'
+import { CATEGORIES_ROUTE, getCategoryPath } from './routes/categories'
 import EventModel from '../endpoint/models/EventModel'
 
 import type { Location } from 'redux-first-router/dist/flow-types'
